requirements.txt
apache-airflow~=2.10.2
polars~=1.8.1